+++ Terminal-Icons/Private/Get-PreferencesFile.ps1
@@ -1,0 +1,8 @@
+function Get-PreferencesFile {
+    [OutputType([String])]
+    param(
+        [string]$Filename = $script:prefsFile
+    )
+
+    [IO.Path]::Combine((Get-ThemeStoragePath), $Filename)
+}