+++ Terminal-Icons/Private/Get-RendorModeUnicode.ps1
@@ -1,0 +1,14 @@
+function Get-RendorModeUnicode {
+    [OutputType([String])]
+    param(
+        [RendorMode]$RendorMode
+    )
+
+    switch ($RendorMode) {
+        ([RendorMode]::Normal)    { $null; break }
+        ([RendorMode]::Bold)      { $script:bold; break}
+        ([RendorMode]::Italic)    { $script:italic; break }
+        ([RendorMode]::Underline) { $script:underline; break }
+        default { $null}
+    }
+}