+++ Terminal-Icons/Private/Import-Preferences.ps1
@@ -1,25 +1,30 @@
 function Import-Preferences {
     [Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSUseShouldProcessForStateChangingFunctions', '')]
-    [OutputType([hashtable])]
+    [Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSUseSingularNouns', '')]
+    [OutputType([Preferences])]
     [cmdletbinding()]
     param(
         [parameter(ValueFromPipeline)]
-        [string]$Path = (Join-Path (Get-ThemeStoragePath) 'prefs.xml'),
-
-        [string]$DefaultThemeName = $script:defaultTheme
+        [string]$Path = (Get-PreferencesFile)
     )
 
     begin {
-        $defaultPrefs = @{
-            CurrentColorTheme = $DefaultThemeName
-            CurrentIconTheme  = $DefaultThemeName
-        }
+        $defaultPrefs = [Preferences]::new()
     }
 
     process {
         if (Test-Path $Path) {
             try {
-                Import-Clixml -Path $Path -ErrorAction Stop
+                $prefs = Import-Clixml -Path $Path -ErrorAction Stop
+                if (-not $prefs -is [Preferences]) {
+                    # Preferences are in old format. Convert to new format.
+                    $newPrefs = [Preferences]::new()
+                    $newPrefs.ColorTheme = $prefs.CurrentColorTheme
+                    $newPrefs.IconTheme  = $prefs.CurrentIconTheme
+                    $newPrefs
+                } else {
+                    $prefs
+                }
             } catch {
                 Write-Warning "Unable to parse [$Path]. Setting default preferences."
                 $defaultPrefs